oops had tsx vibe coded too hard
